Fix the problem that a message box of Address Book does not disappear
This occurs if an user presses Enter key when nothing is selected in a
listbox. Because if Enter key is pressed, both LN_ENTER and DID_OK are
generated. So post a WM_COMMAND message with DID_OK only if a selected
item exists.

Of course, both LN_ENTER and DID_OK are generated even if a selected
item exists. But the first call to DID_OK dismiss a dialog, the second
call does not reach to DID_OK.